overlord: move install handler functions from devicestate to install package
