+++ overlord/install/install.go
@@ -1,0 +1,708 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+
+/*
+ * Copyright (C) 2023 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package install
+
+import (
+	"bytes"
+	"crypto"
+	"encoding/base64"
+	"encoding/hex"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"io"
+	"io/ioutil"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"runtime"
+	"time"
+
+	_ "golang.org/x/crypto/sha3"
+
+	"github.com/snapcore/snapd/asserts"
+	"github.com/snapcore/snapd/asserts/sysdb"
+	"github.com/snapcore/snapd/boot"
+	"github.com/snapcore/snapd/dirs"
+	"github.com/snapcore/snapd/gadget"
+	"github.com/snapcore/snapd/gadget/device"
+	"github.com/snapcore/snapd/kernel/fde"
+	"github.com/snapcore/snapd/logger"
+	"github.com/snapcore/snapd/osutil"
+	"github.com/snapcore/snapd/randutil"
+	"github.com/snapcore/snapd/secboot"
+	"github.com/snapcore/snapd/secboot/keys"
+	"github.com/snapcore/snapd/seed"
+	"github.com/snapcore/snapd/snap"
+	"github.com/snapcore/snapd/snap/squashfs"
+	"github.com/snapcore/snapd/strutil"
+	"github.com/snapcore/snapd/sysconfig"
+	"github.com/snapcore/snapd/timings"
+)
+
+var (
+	secbootTransitionEncryptionKeyChange = secboot.TransitionEncryptionKeyChange
+	sysconfigConfigureTargetSystem       = sysconfig.ConfigureTargetSystem
+)
+
+var timeNow = time.Now
+
+func setSysconfigCloudOptions(opts *sysconfig.Options, gadgetDir string, model *asserts.Model) {
+	ubuntuSeedCloudCfg := filepath.Join(boot.InitramfsUbuntuSeedDir, "data/etc/cloud/cloud.cfg.d")
+
+	grade := model.Grade()
+
+	// we always set the cloud-init src directory if it exists, it is
+	// automatically ignored by sysconfig in the case it shouldn't be used
+	if osutil.IsDirectory(ubuntuSeedCloudCfg) {
+		opts.CloudInitSrcDir = ubuntuSeedCloudCfg
+	}
+
+	switch {
+	// if the gadget has a cloud.conf file, always use that regardless of grade
+	case sysconfig.HasGadgetCloudConf(gadgetDir):
+		opts.AllowCloudInit = true
+
+	// next thing is if are in secured grade and didn't have gadget config, we
+	// disable cloud-init always, clouds should have their own config via
+	// gadgets for grade secured
+	case grade == asserts.ModelSecured:
+		opts.AllowCloudInit = false
+
+	// all other cases we allow cloud-init to run, either through config that is
+	// available at runtime via a CI-DATA USB drive, or via config on
+	// ubuntu-seed if that is allowed by the model grade, etc.
+	default:
+		opts.AllowCloudInit = true
+	}
+}
+
+func writeModel(model *asserts.Model, where string) error {
+	f, err := os.OpenFile(where, os.O_WRONLY|os.O_CREATE|os.O_EXCL, 0644)
+	if err != nil {
+		return err
+	}
+	defer f.Close()
+	return asserts.NewEncoder(f).Encode(model)
+}
+
+func writeTimesyncdClock(srcRootDir, dstRootDir string) error {
+	// keep track of the time
+	const timesyncClockInRoot = "/var/lib/systemd/timesync/clock"
+	clockSrc := filepath.Join(srcRootDir, timesyncClockInRoot)
+	clockDst := filepath.Join(dstRootDir, timesyncClockInRoot)
+	if err := os.MkdirAll(filepath.Dir(clockDst), 0755); err != nil {
+		return fmt.Errorf("cannot store the clock: %v", err)
+	}
+	if !osutil.FileExists(clockSrc) {
+		logger.Noticef("timesyncd clock timestamp %v does not exist", clockSrc)
+		return nil
+	}
+	// clock file is owned by a specific user/group, thus preserve
+	// attributes of the source
+	if err := osutil.CopyFile(clockSrc, clockDst, osutil.CopyFlagPreserveAll); err != nil {
+		return fmt.Errorf("cannot copy clock: %v", err)
+	}
+	// the file is empty however, its modification timestamp is used to set
+	// up the current time
+	if err := os.Chtimes(clockDst, timeNow(), timeNow()); err != nil {
+		return fmt.Errorf("cannot update clock timestamp: %v", err)
+	}
+	return nil
+}
+
+// BuildInstallObserver creates an observer for gadget assets if
+// applicable, otherwise the returned gadget.ContentObserver is nil.
+// The observer if any is also returned as non-nil trustedObserver if
+// encryption is in use.
+func BuildInstallObserver(model *asserts.Model, gadgetDir string, useEncryption bool) (
+	observer gadget.ContentObserver, trustedObserver *boot.TrustedAssetsInstallObserver, err error) {
+
+	// observer will be a nil interface by default
+	trustedObserver, err = boot.TrustedAssetsInstallObserverForModel(model, gadgetDir, useEncryption)
+	if err != nil && err != boot.ErrObserverNotApplicable {
+		return nil, nil, fmt.Errorf("cannot setup asset install observer: %v", err)
+	}
+	if err == nil {
+		observer = trustedObserver
+		if !useEncryption {
+			// there will be no key sealing, so past the
+			// installation pass no other methods need to be called
+			trustedObserver = nil
+		}
+	}
+
+	return observer, trustedObserver, nil
+}
+
+func PrepareEncryptedSystemData(model *asserts.Model, keyForRole map[string]keys.EncryptionKey, trustedInstallObserver *boot.TrustedAssetsInstallObserver) error {
+	// validity check
+	if len(keyForRole) == 0 || keyForRole[gadget.SystemData] == nil || keyForRole[gadget.SystemSave] == nil {
+		return fmt.Errorf("internal error: system encryption keys are unset")
+	}
+	dataEncryptionKey := keyForRole[gadget.SystemData]
+	saveEncryptionKey := keyForRole[gadget.SystemSave]
+
+	// make note of the encryption keys
+	trustedInstallObserver.ChosenEncryptionKeys(dataEncryptionKey, saveEncryptionKey)
+
+	// keep track of recovery assets
+	if err := trustedInstallObserver.ObserveExistingTrustedRecoveryAssets(boot.InitramfsUbuntuSeedDir); err != nil {
+		return fmt.Errorf("cannot observe existing trusted recovery assets: %v", err)
+	}
+	if err := saveKeys(model, keyForRole); err != nil {
+		return err
+	}
+	// write markers containing a secret to pair data and save
+	if err := writeMarkers(model); err != nil {
+		return err
+	}
+	return nil
+}
+
+func PrepareRunSystemData(model *asserts.Model, gadgetDir string, perfTimings timings.Measurer) error {
+	// keep track of the model we installed
+	err := os.MkdirAll(filepath.Join(boot.InitramfsUbuntuBootDir, "device"), 0755)
+	if err != nil {
+		return fmt.Errorf("cannot store the model: %v", err)
+	}
+	err = writeModel(model, filepath.Join(boot.InitramfsUbuntuBootDir, "device/model"))
+	if err != nil {
+		return fmt.Errorf("cannot store the model: %v", err)
+	}
+
+	// preserve systemd-timesyncd clock timestamp, so that RTC-less devices
+	// can start with a more recent time on the next boot
+	if err := writeTimesyncdClock(dirs.GlobalRootDir, boot.InstallHostWritableDir(model)); err != nil {
+		return fmt.Errorf("cannot seed timesyncd clock: %v", err)
+	}
+
+	// configure the run system
+	opts := &sysconfig.Options{TargetRootDir: boot.InstallHostWritableDir(model), GadgetDir: gadgetDir}
+	// configure cloud init
+	setSysconfigCloudOptions(opts, gadgetDir, model)
+	// allow running this helper without perfTimings
+	if perfTimings == nil {
+		err = sysconfigConfigureTargetSystem(model, opts)
+	} else {
+		timings.Run(perfTimings, "sysconfig-configure-target-system", "Configure target system", func(timings.Measurer) {
+			err = sysconfigConfigureTargetSystem(model, opts)
+		})
+	}
+	if err != nil {
+		return err
+	}
+
+	// TODO: FIXME: this should go away after we have time to design a proper
+	//              solution
+
+	if !model.Classic() {
+		// on some specific devices, we need to create these directories in
+		// _writable_defaults in order to allow the install-device hook to install
+		// some files there, this eventually will go away when we introduce a proper
+		// mechanism not using system-files to install files onto the root
+		// filesystem from the install-device hook
+		if err := fixupWritableDefaultDirs(boot.InstallHostWritableDir(model)); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func fixupWritableDefaultDirs(systemDataDir string) error {
+	// the _writable_default directory is used to put files in place on
+	// ubuntu-data from install mode, so we abuse it here for a specific device
+	// to let that device install files with system-files and the install-device
+	// hook
+
+	// eventually this will be a proper, supported, designed mechanism instead
+	// of just this hack, but this hack is just creating the directories, since
+	// the system-files interface only allows creating the file, not creating
+	// the directories leading up to that file, and since the file is deeply
+	// nested we would effectively have to give all permission to the device
+	// to create any file on ubuntu-data which we don't want to do, so we keep
+	// this restriction to let the device create one specific file, and then
+	// we behind the scenes just create the directories for the device
+
+	for _, subDirToCreate := range []string{"/etc/udev/rules.d", "/etc/modprobe.d", "/etc/modules-load.d/", "/etc/systemd/network"} {
+		dirToCreate := sysconfig.WritableDefaultsDir(systemDataDir, subDirToCreate)
+
+		if err := os.MkdirAll(dirToCreate, 0755); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// writeMarkers writes markers containing the same secret to pair data and save.
+func writeMarkers(model *asserts.Model) error {
+	// ensure directory for markers exists
+	if err := os.MkdirAll(boot.InstallHostFDEDataDir(model), 0755); err != nil {
+		return err
+	}
+	if err := os.MkdirAll(boot.InstallHostFDESaveDir, 0755); err != nil {
+		return err
+	}
+
+	// generate a secret random marker
+	markerSecret, err := randutil.CryptoTokenBytes(32)
+	if err != nil {
+		return fmt.Errorf("cannot create ubuntu-data/save marker secret: %v", err)
+	}
+
+	return device.WriteEncryptionMarkers(boot.InstallHostFDEDataDir(model), boot.InstallHostFDESaveDir, markerSecret)
+}
+
+func saveKeys(model *asserts.Model, keyForRole map[string]keys.EncryptionKey) error {
+	saveEncryptionKey := keyForRole[gadget.SystemSave]
+	if saveEncryptionKey == nil {
+		// no system-save support
+		return nil
+	}
+	// ensure directory for keys exists
+	if err := os.MkdirAll(boot.InstallHostFDEDataDir(model), 0755); err != nil {
+		return err
+	}
+	if err := saveEncryptionKey.Save(device.SaveKeyUnder(boot.InstallHostFDEDataDir(model))); err != nil {
+		return fmt.Errorf("cannot store system save key: %v", err)
+	}
+	return nil
+}
+
+func ReadPreseedAssertion(assertDb *asserts.Database, model *asserts.Model, ubuntuSeedDir, sysLabel string) (*asserts.Preseed, error) {
+	f, err := os.Open(filepath.Join(ubuntuSeedDir, "systems", sysLabel, "preseed"))
+	if err != nil {
+		return nil, fmt.Errorf("cannot read preseed assertion: %v", err)
+	}
+
+	batch := asserts.NewBatch(nil)
+	_, err = batch.AddStream(f)
+	if err != nil {
+		return nil, err
+	}
+
+	var preseedAs *asserts.Preseed
+	err = batch.CommitToAndObserve(assertDb, func(as asserts.Assertion) {
+		if as.Type() == asserts.PreseedType {
+			preseedAs = as.(*asserts.Preseed)
+		}
+	}, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	switch {
+	case preseedAs == nil:
+		return nil, fmt.Errorf("internal error: preseed assertion file is present but preseed assertion not found")
+	case preseedAs.SystemLabel() != sysLabel:
+		return nil, fmt.Errorf("preseed assertion system label %q doesn't match system label %q", preseedAs.SystemLabel(), sysLabel)
+	case preseedAs.Model() != model.Model():
+		return nil, fmt.Errorf("preseed assertion model %q doesn't match the model %q", preseedAs.Model(), model.Model())
+	case preseedAs.BrandID() != model.BrandID():
+		return nil, fmt.Errorf("preseed assertion brand %q doesn't match model brand %q", preseedAs.BrandID(), model.BrandID())
+	case preseedAs.Series() != model.Series():
+		return nil, fmt.Errorf("preseed assertion series %q doesn't match model series %q", preseedAs.Series(), model.Series())
+	}
+
+	return preseedAs, nil
+}
+
+var seedOpen = seed.Open
+
+// TODO: consider reusing this kind of handler for UC20 seeding
+type preseedSnapHandler struct {
+	writableDir string
+}
+
+func (p *preseedSnapHandler) HandleUnassertedSnap(name, path string, _ timings.Measurer) (string, error) {
+	pinfo := snap.MinimalPlaceInfo(name, snap.Revision{N: -1})
+	targetPath := filepath.Join(p.writableDir, pinfo.MountFile())
+	mountDir := filepath.Join(p.writableDir, pinfo.MountDir())
+
+	sq := squashfs.New(path)
+	opts := &snap.InstallOptions{MustNotCrossDevices: true}
+	if _, err := sq.Install(targetPath, mountDir, opts); err != nil {
+		return "", fmt.Errorf("cannot install snap %q: %v", name, err)
+	}
+
+	return targetPath, nil
+}
+
+func (p *preseedSnapHandler) HandleAndDigestAssertedSnap(name, path string, essType snap.Type, snapRev *asserts.SnapRevision, _ func(string, uint64) (snap.Revision, error), _ timings.Measurer) (string, string, uint64, error) {
+	pinfo := snap.MinimalPlaceInfo(name, snap.Revision{N: snapRev.SnapRevision()})
+	targetPath := filepath.Join(p.writableDir, pinfo.MountFile())
+	mountDir := filepath.Join(p.writableDir, pinfo.MountDir())
+
+	logger.Debugf("copying: %q to %q; mount dir=%q", path, targetPath, mountDir)
+
+	srcFile, err := os.Open(path)
+	if err != nil {
+		return "", "", 0, err
+	}
+	defer srcFile.Close()
+
+	destFile, err := osutil.NewAtomicFile(targetPath, 0644, 0, osutil.NoChown, osutil.NoChown)
+	if err != nil {
+		return "", "", 0, fmt.Errorf("cannot create atomic file: %v", err)
+	}
+	defer destFile.Cancel()
+
+	finfo, err := srcFile.Stat()
+	if err != nil {
+		return "", "", 0, err
+	}
+
+	destFile.SetModTime(finfo.ModTime())
+
+	h := crypto.SHA3_384.New()
+	w := io.MultiWriter(h, destFile)
+
+	size, err := io.CopyBuffer(w, srcFile, make([]byte, 2*1024*1024))
+	if err != nil {
+		return "", "", 0, err
+	}
+	if err := destFile.Commit(); err != nil {
+		return "", "", 0, fmt.Errorf("cannot copy snap %q: %v", name, err)
+	}
+
+	sq := squashfs.New(targetPath)
+	opts := &snap.InstallOptions{MustNotCrossDevices: true}
+	// since Install target path is the same as source path passed to squashfs.New,
+	// Install isn't going to copy the blob, but we call it to set up mount directory etc.
+	if _, err := sq.Install(targetPath, mountDir, opts); err != nil {
+		return "", "", 0, fmt.Errorf("cannot install snap %q: %v", name, err)
+	}
+
+	sha3_384, err := asserts.EncodeDigest(crypto.SHA3_384, h.Sum(nil))
+	if err != nil {
+		return "", "", 0, fmt.Errorf("cannot encode snap %q digest: %v", path, err)
+	}
+	return targetPath, sha3_384, uint64(size), nil
+}
+
+func MaybeApplyPreseededData(model *asserts.Model, preseedAs *asserts.Preseed, preseedArtifact, ubuntuSeedDir, sysLabel, writableDir string) error {
+
+	// TODO: consider a writer that feeds the file to stdin of tar and calculates the digest at the same time.
+	sha3_384, _, err := osutil.FileDigest(preseedArtifact, crypto.SHA3_384)
+	if err != nil {
+		return fmt.Errorf("cannot calculate preseed artifact digest: %v", err)
+	}
+
+	digest, err := base64.RawURLEncoding.DecodeString(preseedAs.ArtifactSHA3_384())
+	if err != nil {
+		return fmt.Errorf("cannot decode preseed artifact digest")
+	}
+	if !bytes.Equal(sha3_384, digest) {
+		return fmt.Errorf("invalid preseed artifact digest")
+	}
+
+	logger.Noticef("apply preseed data: %q, %q", writableDir, preseedArtifact)
+	cmd := exec.Command("tar", "--extract", "--preserve-permissions", "--preserve-order", "--gunzip", "--directory", writableDir, "-f", preseedArtifact)
+	if err := cmd.Run(); err != nil {
+		return err
+	}
+
+	logger.Noticef("copying snaps")
+
+	deviceSeed, err := seedOpen(ubuntuSeedDir, sysLabel)
+	if err != nil {
+		return err
+	}
+	tm := timings.New(nil)
+
+	if err := deviceSeed.LoadAssertions(nil, nil); err != nil {
+		return err
+	}
+
+	if err := os.MkdirAll(filepath.Join(writableDir, "var/lib/snapd/snaps"), 0755); err != nil {
+		return err
+	}
+
+	n := runtime.NumCPU()
+	if n > 1 {
+		deviceSeed.SetParallelism(n)
+	}
+	snapHandler := &preseedSnapHandler{writableDir: writableDir}
+	if err := deviceSeed.LoadMeta("run", snapHandler, tm); err != nil {
+		return err
+	}
+
+	preseedSnaps := make(map[string]*asserts.PreseedSnap)
+	for _, ps := range preseedAs.Snaps() {
+		preseedSnaps[ps.Name] = ps
+	}
+
+	checkSnap := func(ssnap *seed.Snap) error {
+		ps, ok := preseedSnaps[ssnap.SnapName()]
+		if !ok {
+			return fmt.Errorf("snap %q not present in the preseed assertion", ssnap.SnapName())
+		}
+		if ps.Revision != ssnap.SideInfo.Revision.N {
+			rev := snap.Revision{N: ps.Revision}
+			return fmt.Errorf("snap %q has wrong revision %s (expected: %s)", ssnap.SnapName(), ssnap.SideInfo.Revision, rev)
+		}
+		if ps.SnapID != ssnap.SideInfo.SnapID {
+			return fmt.Errorf("snap %q has wrong snap id %q (expected: %q)", ssnap.SnapName(), ssnap.SideInfo.SnapID, ps.SnapID)
+		}
+		return nil
+	}
+
+	esnaps := deviceSeed.EssentialSnaps()
+	msnaps, err := deviceSeed.ModeSnaps("run")
+	if err != nil {
+		return err
+	}
+	if len(msnaps)+len(esnaps) != len(preseedSnaps) {
+		return fmt.Errorf("seed has %d snaps but %d snaps are required by preseed assertion", len(msnaps)+len(esnaps), len(preseedSnaps))
+	}
+
+	for _, esnap := range esnaps {
+		if err := checkSnap(esnap); err != nil {
+			return err
+		}
+	}
+
+	for _, ssnap := range msnaps {
+		if err := checkSnap(ssnap); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func RestoreDeviceFromSave(model *asserts.Model) error {
+	// we could also look at factory-reset-bootstrap.json left by
+	// snap-bootstrap, but the mount was already verified during boot
+	mounted, err := osutil.IsMounted(boot.InitramfsUbuntuSaveDir)
+	if err != nil {
+		return fmt.Errorf("cannot determine ubuntu-save mount state: %v", err)
+	}
+	if !mounted {
+		logger.Noticef("not restoring from save, ubuntu-save not mounted")
+		return nil
+	}
+	// TODO anything else we want to restore?
+	return RestoreDeviceSerialFromSave(model)
+}
+
+func RestoreDeviceSerialFromSave(model *asserts.Model) error {
+	fromDevice := filepath.Join(boot.InstallHostDeviceSaveDir)
+	logger.Debugf("looking for serial assertion and device key under %v", fromDevice)
+	fromDB, err := sysdb.OpenAt(fromDevice)
+	if err != nil {
+		return err
+	}
+	// key pair manager always uses ubuntu-save whenever it's available
+	kp, err := asserts.OpenFSKeypairManager(fromDevice)
+	if err != nil {
+		return err
+	}
+	// there should be a serial assertion for the current model
+	serials, err := fromDB.FindMany(asserts.SerialType, map[string]string{
+		"brand-id": model.BrandID(),
+		"model":    model.Model(),
+	})
+	if (err != nil && errors.Is(err, &asserts.NotFoundError{})) || len(serials) == 0 {
+		// there is no serial assertion in the old system that matches
+		// our model, it is still possible that the old system could
+		// have generated device keys and sent out a serial request, but
+		// for simplicity we ignore this scenario and a new set of keys
+		// will be generated after booting into the run system
+		logger.Debugf("no serial assertion for %v/%v", model.BrandID(), model.Model())
+		return nil
+	}
+	if err != nil {
+		return err
+	}
+	logger.Noticef("found %v serial assertions for %v/%v", len(serials), model.BrandID(), model.Model())
+
+	var serialAs *asserts.Serial
+	for _, serial := range serials {
+		maybeCurrentSerialAs := serial.(*asserts.Serial)
+		// serial assertion is signed with the device key, its ID is in the
+		// header
+		deviceKeyID := maybeCurrentSerialAs.DeviceKey().ID()
+		logger.Debugf("serial assertion device key ID: %v", deviceKeyID)
+
+		// there can be multiple serial assertions, as the device could
+		// have exercised the registration a number of times, but each
+		// time it unregisters, the old key is removed and a new one is
+		// generated
+		_, err = kp.Get(deviceKeyID)
+		if err != nil {
+			if asserts.IsKeyNotFound(err) {
+				logger.Debugf("no key with ID %v", deviceKeyID)
+				continue
+			}
+			return fmt.Errorf("cannot obtain device key: %v", err)
+		} else {
+			serialAs = maybeCurrentSerialAs
+			break
+		}
+	}
+
+	if serialAs == nil {
+		// no serial assertion that matches the model, brand and is
+		// signed with a device key that is present in the filesystem
+		logger.Debugf("no valid serial assertions")
+		return nil
+	}
+
+	logger.Debugf("found a serial assertion for %v/%v, with serial %v",
+		model.BrandID(), model.Model(), serialAs.Serial())
+
+	toDB, err := sysdb.OpenAt(filepath.Join(boot.InstallHostWritableDir(model), "var/lib/snapd/assertions"))
+	if err != nil {
+		return err
+	}
+
+	logger.Debugf("importing serial and model assertions")
+	b := asserts.NewBatch(nil)
+	err = b.Fetch(toDB,
+		func(ref *asserts.Ref) (asserts.Assertion, error) { return ref.Resolve(fromDB.Find) },
+		func(f asserts.Fetcher) error {
+			if err := f.Save(model); err != nil {
+				return err
+			}
+			return f.Save(serialAs)
+		})
+	if err != nil {
+		return fmt.Errorf("cannot fetch assertions: %v", err)
+	}
+	if err := b.CommitTo(toDB, nil); err != nil {
+		return fmt.Errorf("cannot commit assertions: %v", err)
+	}
+	return nil
+}
+
+type factoryResetMarker struct {
+	FallbackSaveKeyHash string `json:"fallback-save-key-sha3-384,omitempty"`
+}
+
+func fileDigest(p string) (string, error) {
+	digest, _, err := osutil.FileDigest(p, crypto.SHA3_384)
+	if err != nil {
+		return "", err
+	}
+	return hex.EncodeToString(digest), nil
+}
+
+func WriteFactoryResetMarker(marker string, hasEncryption bool) error {
+	keyDigest := ""
+	if hasEncryption {
+		d, err := fileDigest(device.FactoryResetFallbackSaveSealedKeyUnder(boot.InitramfsSeedEncryptionKeyDir))
+		if err != nil {
+			return err
+		}
+		keyDigest = d
+	}
+	var buf bytes.Buffer
+	err := json.NewEncoder(&buf).Encode(factoryResetMarker{
+		FallbackSaveKeyHash: keyDigest,
+	})
+	if err != nil {
+		return err
+	}
+
+	if hasEncryption {
+		logger.Noticef("writing factory-reset marker at %v with key digest %q", marker, keyDigest)
+	} else {
+		logger.Noticef("writing factory-reset marker at %v", marker)
+	}
+	return osutil.AtomicWriteFile(marker, buf.Bytes(), 0644, 0)
+}
+
+func VerifyFactoryResetMarkerInRun(marker string, hasEncryption bool) error {
+	f, err := os.Open(marker)
+	if err != nil {
+		return err
+	}
+	defer f.Close()
+	var frm factoryResetMarker
+	if err := json.NewDecoder(f).Decode(&frm); err != nil {
+		return err
+	}
+	if hasEncryption {
+		saveFallbackKeyFactory := device.FactoryResetFallbackSaveSealedKeyUnder(boot.InitramfsSeedEncryptionKeyDir)
+		d, err := fileDigest(saveFallbackKeyFactory)
+		if err != nil {
+			// possible that there was unexpected reboot
+			// before, after the key was moved, but before
+			// the marker was removed, in which case the
+			// actual fallback key should have the right
+			// digest
+			if !os.IsNotExist(err) {
+				// unless it's a different error
+				return err
+			}
+			saveFallbackKeyFactory := device.FallbackSaveSealedKeyUnder(boot.InitramfsSeedEncryptionKeyDir)
+			d, err = fileDigest(saveFallbackKeyFactory)
+			if err != nil {
+				return err
+			}
+		}
+		if d != frm.FallbackSaveKeyHash {
+			return fmt.Errorf("fallback sealed key digest mismatch, got %v expected %v", d, frm.FallbackSaveKeyHash)
+		}
+	} else {
+		if frm.FallbackSaveKeyHash != "" {
+			return fmt.Errorf("unexpected non-empty fallback key digest")
+		}
+	}
+	return nil
+}
+
+func RotateEncryptionKeys() error {
+	kd, err := ioutil.ReadFile(filepath.Join(dirs.SnapFDEDir, "ubuntu-save.key"))
+	if err != nil {
+		return fmt.Errorf("cannot open encryption key file: %v", err)
+	}
+	// does the right thing if the key has already been transitioned
+	if err := secbootTransitionEncryptionKeyChange(boot.InitramfsUbuntuSaveDir, keys.EncryptionKey(kd)); err != nil {
+		return fmt.Errorf("cannot transition the encryption key: %v", err)
+	}
+	return nil
+}
+
+func checkFDEFeatures(runSetupHook fde.RunSetupHookFunc) (et secboot.EncryptionType, err error) {
+	// Run fde-setup hook with "op":"features". If the hook
+	// returns any {"features":[...]} reply we consider the
+	// hardware supported. If the hook errors or if it returns
+	// {"error":"hardware-unsupported"} we don't.
+	features, err := fde.CheckFeatures(runSetupHook)
+	if err != nil {
+		return et, err
+	}
+	if strutil.ListContains(features, "device-setup") {
+		et = secboot.EncryptionTypeDeviceSetupHook
+	} else {
+		et = secboot.EncryptionTypeLUKS
+	}
+
+	return et, nil
+}
+
+func HasFDESetupHookInKernel(kernelInfo *snap.Info) bool {
+	_, ok := kernelInfo.Hooks["fde-setup"]
+	return ok
+}
